+++ rcos_io/blueprints/projects/templates/projects/detail.html
